requirements.txt
playwright>=1.40.0
ddddocr>=1.4.11
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
